+++ lib/hmmlearn/tests/test_multinomial_hmm.py
@@ -116,10 +116,14 @@ class TestMultinomailHMM:
 
         assert log_likelihood_increasing(h, X, lengths, n_iter)
 
-    def test__check_and_set_n_features(self):
-        self.h._check_and_set_n_features(np.array([[0, 0, 2, 1, 3, 1, 1]]))
-        self.h._check_and_set_n_features(np.array([[0, 0, 1, 3, 1]], np.uint8))
+    def test__check_and_set_multinomial_n_features(self):
+        self.h._check_and_set_multinomial_n_features(
+            np.array([[0, 0, 2, 1, 3, 1, 1]]))
+        self.h._check_and_set_multinomial_n_features(
+            np.array([[0, 0, 1, 3, 1]], np.uint8))
         with pytest.raises(ValueError):  # non-integral
-            self.h._check_and_set_n_features(np.array([[0., 2., 1., 3.]]))
+            self.h._check_and_set_multinomial_n_features(
+                np.array([[0., 2., 1., 3.]]))
         with pytest.raises(ValueError):  # negative integers
-            self.h._check_and_set_n_features(np.array([[0, -2, 1, 3, 1, 1]]))
+            self.h._check_and_set_multinomial_n_features(
+                np.array([[0, -2, 1, 3, 1, 1]]))